requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-core
langchain-google-genai
google-generativeai
sentence-transformers
chromadb
pypdf
python-dotenv